Renaming cardColor to cardData to fit purpose
